requirements.txt
certifi==2018.11.29
chardet==3.0.4
idna==2.7
python-dateutil==2.7.5
requests==2.20.1
six==1.11.0
urllib3==1.24.1
xlrd==1.1.0
